Changing sort from int to double

+++ lib/features/products/data/datasources/remote/products_remote_datasource.dart
@@ -40,8 +40,8 @@ class ProductsRemoteDatasource extends DioDatasource {
       // Resulto ser porque el punto lo tomaba como un punto flotante, tomando solo 2 decimales,
       // Al quitar el punto del string ya se soluciono el problema
       results.sort((a, b) {
-        final priceA = int.parse(a.price.replaceAll("\.", ""));
-        final priceB = int.parse(b.price.replaceAll("\.", ""));
+        final priceA = double.parse(a.price.replaceAll("\.", ""));
+        final priceB = double.parse(b.price.replaceAll("\.", ""));
 
         return priceA.compareTo(priceB);
       });
